+++ test/test_helper.rb
@@ -9,8 +9,9 @@ require 'minitest/reporters'
 require 'minitest'
 
 $VERBOSE = nil
+Appium::Logger.level = ::Logger::FATAL # Show Logger logs only they are fatal
 
-Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
+Minitest::Reporters.use! Minitest::Reporters::ProgressReporter.new
 
 ROOT_REPORT_PATH = "#{Dir.pwd}/test/report".freeze
 START_AT = Time.now.strftime('%Y-%m-%d-%H%M%S').freeze